Create section to keep html content for all modals in one place
The .modals section contains the dialog element for task add modal.

Since the "Add Task" button will be located in multiple places on
the webpage, the same modal will be called for all of them, hence
the reason for a separate section.

The section will also contain the dialog for displaying task in
full view, the dialog for add projects, etc.