Create module for object constructors for the DOM
The module will store DOM-side object constructors which are called
in another module to create new instances of the DOM content of the
task, project, etc.

The TaskDOM object constructor function creates a DOM element for
the Task that is created and returns it. A unique ID is attached
to the DOM element for later-use.

The taskListDOM array is exported alongside the TaskDOM object
constructor function in order to store the DOM element of each newly
created task for performing task-specific actions.

The object creation pattern currently being used is the object
constructor functions approach. This might be replaced by Classes
or Factory Functions later as per need.