Create module for creating instances of objects upon user input
The module imports the object constructor functions and their
respective arrays for application-side and DOM-side.

The idGen variable is used to set unique ID of each newly created
task.

The taskConstructor() function sets up event listeners on task-add
buttons and displays the modal for adding task. The same function
also sets up event listener on the submit button of the modal.

Inside the event listener of the submit button, an instance of the
Task object (application-side) and TaskDOM object (DOM-side) is
created and logged to console.
The id generator variable (idGen) is also incremented and the modal
is closed.

The taskConstructor() function is exported for use in index.js