Replace DOM object constructor for task with a modal to display task
Initially, the DOM content of task was created dynamically using
an object constructor.
This has been replaced with a modal instead. When displayed, the
task-specific values will be passed to the modal and the task will
be displayed.

The DOM content object constructor will instead be used for superset
like project, workspace, stc.