Refactor ProjectDOM factory function
Initially, the ProjectDOM factory function returned methods to
update DOM content as part of an object.

However, there was need to include methods to add to list or remove
from list.
Since these methods are also to be used elsewhere, they will be
defined as separate methods inside an object and will be assigned
to the returned value of ProjectDOM using Object.assign().
In essence, we are doing this to take advantage of composition.

Now, the ProjectDOM function returns a new object to which the
object updating DOM content are assigned.