Refactor task creation code and setup project property of task object
The taskList array has been scrapped and instead, the task is now
stored inside the array of the parent project object.

The parent project (project of which the task is a part of) is given
as input by the user and the task is pushed inside the array of that
project by traversing the projectList array.