Instantiate the inbox object; change id generator initial value
Inbox serves as the default project. In case the user does not
want the task to be part of any project, the default destination
is the inbox.

The inbox has been setup as an IIFE since it will only ever be a
single instance.

The id generator value has been initialised to 1 since the Inbox
instance will be carrying the id of 0.