Refactor Project object constructor function and assign list methods
The list methods imported from methods module consist of list
operation methods. These methods are useful in case of Project
object as they carry an array of tasks.

Therefore, we have refactored the Project object constructor function
in order to return an object consisting of the properties of both
the listMethods() factory as well as the basic Project object we
initially had using Object.assign().