Refactor drop-down list generator method
The <option> elements which are dynamically created initially had
the name of item as both the value and the text content.
However, the value attribute will now contain the id of the item
and the text content will have the name.

This will aid us in applying list methods to the dynamically created
elements.