Refactor remove() and search() list methods
The methods initially took in the whole object as parameter. This
created issues when we needed to apply these operations but could
not for some reason send the whole object but could only send the
id (which is actually used in code).

Hence, the methods now take in the id of the target.