Replace projectList global array with an IIFE with list methods
The projectList array was in global scope. Hence, to reduce global
clutter, we have replaced it with the ProjectList factory IIFE.

The IIFE returns an object having a list as well as the list methods
defined in methods module.

As a result, all the instances of the uses of ProjectList have been
updated as well.