Replace projectListDOM global array with an IIFE with list methods
The projectListDOM array was in global scope. Hence, to reduce global
clutter, we have replaced it with the ProjectListDOM factory IIFE.

The IIFE returns an object having a list as well as the list methods
defined in methods module.

As a result, all the instances of the uses of ProjectListDOM have been
updated as well.