Create IIFE with display control methods
The displayControl() IIFE has two purposes:
1. To update the task list of each project's main display.
2. To update the display in case the project's main display is
already active. This is useful when we have to dynamically update
the display while adding task to the project being displayed.

The IIFE is used in the project DOM object constructor function to
update the display with list of tasks.

The IIFE is used in the event when a task is created and the parent
project is being displayed in order to update the display right after
the task is successfully added.