Trigger event on task-add btn; use displayControl method
The "Add Task" button of the project main display dispatches a custom
event to the main "Add Task" button located on sidebar. This saves
us from writing long lines of code again & again.

The displayControl() IIFE method to update the task list is used
in order to update the main section display of the project.