Refactor addEventListener() code for task/project add buttons
Initially, we used the querySelectorALL() method to store all the
instances of add button in an array and applied event listener
to each of them by traversing through the array.

Now, we only attach event listener to the add button located in
sidebar and for any add button located elsewhere, we just dispatch
the event to the sidebar add button. This saves a lot of lines of
code and meaningless traersing of an array.
Also, the array traversal didn't work for any newly added projects
or tasks.
Hence, dispatch event is the best solution in this case.