Refactor listMethod search & remove code
Both search and remove methods made use of binary search. Therefore,
binary search algorithm has been created seperately serving as the
common go-to function for both the methods that use it.

This saves us from writing the same binary search code twice. We
just take in a boolean serving as whether we want to search or we
want to remove and perform the operation accordingly.