Add method to show full task window; setup listener for the same
The taskDisplay() method of displayControl() IIFE displays the task
window/modal that shows the full task specifications.

The event listener of displaying the task window has been set up on
the task DOM element created in updateProjectTaskList() and the
element to be clicked has been changed into a <button> element.