Setup TaskList IIFE for storing all the tasks, similar to ProjectList
The TaskList IIFE will store all the tasks regardless of their parent
project. This is needed when we need to search for a task with only
a task-id and nothing else.
In such a case in the absence of TaskList, we'd have to search for
the task one by one in each project object which can be a lengthy
process.

Therefore, for situations precisely as this, we make use of TaskList
to search for a task among all the tasks.