Rename taskConstructor() function and turn into IIFE; rename add()
The add() function which was situated inside the taskConstrictor()
function has been renamed to taskConstructor instead and the
taskConstructor() function which initially existed as a function
has been converted into an IIFE of the name TaskControl.

To sum up, the parent function is now an IIFE named TaskControl and
the function which handles task construction has been named
taskConstructor.

This has been done so that the task editing functionality can be
added later on alongside the taskConstructor() inside the IIFE
TaskControl which serves as our task control unit.