Update display order for sidebar project tabs
Initially, each project was added before the "Add Project" button
in the sidebar.
However, in case the number of projects on display is a lot, it
would cause clutter on the sidebar display and the "Add Project"
button would be shifted way at the bottom.

Therefore, the sidebarDisplay() method has been updated so that the
projects are added to display after the add button.