Fix search query in refreshProjectDisplay()
Initially the search query used to search for the project within
the list using the ID of project had the argument "task.project".

However, there was no task object present inside the function.
Therefore, the correct argument has been entered, that is, the
projectID variable which actually stores the id of project to be
searched.