Prevent refreshProjectDisplay() from opening project tab
Initially, each time the function was called, it would open the
project tab regardless of whether it was open or not.
This however, was not desired as the function is needed in situations
where the project tab is already and display and we need to update
it with the task added.

Therefore, now, we store the child element of main section in a
variable and check whether a project is in display using the id
attribute. Hence, the function only works in case a project is in
display.