Refactor & rename refreshProjectDisplay() to refresh other tabs
The refreshProjectDisplay() function was used only to refresh the
projects tab if they were in display.

However, we needed similar function for refreshing the "My Projects"
page display in case of an update.

Therefore, we have renamed the function to refreshDisplay() and made
the projectID parameter a default one with value of null in case it
is not provided (in case of "My Projects" tab).
We check the class attribute in case of "My Projects" tab and the
id attribute in case of projects.