Fix refreshDisplay() conditional to prevent unwanted activation
The refreshDisplay() is called each time a project is created.
In case the "My Projects" tab is open, the function works fine.

However, in case "My Projects" tab is not open, then the id
attribute of the mainChild variable would be undefined and the
projectID would be null.
This creates a bug as in conditionals, undefined and null return
true in soft-equality check.

The bug has been fixed by added another conditional to only allow
activation if the projectID is not null.