Modify setupTaskEditFields() for use in other module
The setupTaskEditFields() function is necessary for initiating the
task edit-process.

Now that the task edit function is available on project DOM content
as well, we require this function in order to enable editing for
the DOM in addition to the modal.

As a result, the function has been moved outside the taskEditor()
so that it can be exported for use.