Move the code for deleting task in a separate function for export
Initially, the taskDestructor() function set event listener on the
delete button located in the task display modal. The event listener
hosted the code to delete the task.

However, now that we have delete buttons for each task in the DOM
content of projects, we need to use the same code inside the event
listeners of the delete buttons of each task in project DOM content.

Therefore, the code for deleting the task has been moved into
deleteTask() function and used as required.