Replace ProjectDisplayControl IIFE with simplified projectDisplay()
The IIFE took a roundabout approach to fetch the id of the project
tab being clicked and then generate the DOM content of the project.

This, however, has been avoided by attaching the event listener in
the constructor itself. This leaves room only for the code that
displays the project DOM content.

Therefore, displayProject() function takes in the project as
argument and displays that project.
This saves us a lot of needless code and simplifies the process.