Create edit & delete buttons for sidebar project tabs
The project tab in sidebar has been rid of its child elements which
were used as icon and text of project. Instead, we'll be depending
on list-style design for icon and textContent of the button for text.

The project tab button has been moved inside a list-item and is now
a child of the same along with edit and delete buttons.