Update refreshDisplay() to deal with deleted projects
When projects are deleted, the process at some point calls the
refreshDisplay() function.
However, when a project on display was deleted, it caused the
refreshDisplay() function to break as the project no longer existed.

Therefore, the function has been updated to deal with such cases by
checking if the project exists in the project-lists.
In case the project does not exist in project-lists, it means that
the project has been deleted.
In such a case if the id is provided, the function simply clears the
main display.