Refactor refreshDisplay() function & update conditionals
The conditional which checked whether a deleted project is being
displayed has been moved inside the conditional dealing with
projects. That place is better suited as now the conditional is
placed alongside the conditionals handling the display of projects.

The primary conditionals that checked whether the "My Projects" tab
is being displayed or a project is being displayed has been updated
with checks for whether the mainChild node (the direct child node of
main element section) is valid or not.
This has been done because when the child node of main section is
deleted in case of the display of a deleted project, it lead to
conflicts when the refreshDisplay() function was further called as
the mainChild node now had the value 'null' which gave errors
when the code in conditionals was executed.
Hence the changes made provide us with a safety-check and prevent
any arise of errors.