Remove footer element and update CSS Grid settings
The <footer> element was initially to be used to show the GitHub
link.
However, since the <main> element has too much free space, the link
will be displayed there instead. It is also not confirmed if the
link is even needed to be displayed. Hence, the element has been
removed and the Grid has been updated.