Attach task id to edit modal submit btn to aid task search for editing
Initially, the submit btn event listener of task edit modal searched
the task by reading the id attribute of the edit btn of the task
display modal.
This worked fine as long the the task edit was initiated from the
display modal itself.
However, in case task was edited from the Home page or the project
page, the task display modal had no role to play and had no id
attached to it. This lead to errors in editing process of tasks from
Home page or project page.

Therefore, the setupTaskEditFields() function has been updated so that
the submit btn of the edit modal has the task id attached to it. This
is because no matter where the editing is initiated, the edit modal
is bound to be opened and hence, there will be no issues in fetching
the task id.

The event listener of the submit btn has also been updated so that the
task is searched using the id of the submit btn of edit modal which is
always in use rather than the id of the edit btn of display modal which
is only used when task editing is initiated from the display modal
itself.