Pass the task itself as the argument in task edit event listener
Initially, the task was first searched for in the task list and then
passed off as an argument to the setupTaskEditFields() function of
the task edit event listener of updateProjectTaskList() function.

This however was completely redundant as we already had the task in
scope and we could directly pass it off as an argument.