Move function to setup project edit modal to the constructor module
The function which was known as editProject() in the
dom-object-constructors module has been moved to the ProjectControl
IIFE of constructor module where it is supposed to be situated.

Any code handling the construction, maintenance, and destruction of
an entity is stored in the constructor module.
The other module only stores the code related to generation of the
DOM content of an entity.