Add feature to store data using Web Storage API
The "local-storage.js" module utilizes the localStorage function of
Web Storage API in order to store the To-Dos creates by user. This
makes sure that users do not lose their data upon closing the app or
reloading the page.

The module consists of functions to check whether localStorage is
available, creating a JSON friendly substitute object for our tasks
and projects, functions to retrieve data from storage and store data
in storage, etc.