Refactor projectDestructor() and create new delete call method
The projectDestructor() now sets up the event listeners in project
delete confirmation modal buttons with the code to delete a project
and all its tasks.

The confirmDelete() method will now be called upon the click of a
delete button instead of projectDestructor() and will show the delete
confirmation modal.

The projectDestructor() method sets up the event listeners and is
called in the index module.